Add assets:precompile task
Because we don't wish to keep compiled CSS in git, we need a way to
compile these on deploy. This task tries to find every config.rb task in
your folders, and cleans and compiles the compass files, and then
uploads it to the respective path in the current working dir.

Possible options:

* compass_folders: an array of compass folders. Autodetected if not set
* asset_cache_dir: the directory to store the cache. Defaults to PWD + cache/cached_assets
* compass_output: the way the css file output should be. Defaults to compressed

At the moment this is not added by default. If you want asset
compilation on deploy, add the following to your Capfile

    after 'deploy:update_code', 'assets:precompile'